BUG Create DType type_lists, and extended variant. Constraint ArrayElementProxy conversion operators to type_list. Switch dispatch to use type_list to avoid missing switch cases.

+++ tests/cxx/test_operations.cc
@@ -66,7 +66,7 @@ TEST(NCArrayOperationsTest, Comparisons) {
   ncarray::NCArray arr2(shape, ncarray::DType::float32);
 
   for (ssize_t i = 0; i < 5; ++i) {
-    arr1[{i}] = i;
+    arr1[{i}] = static_cast<float>(i);
   }
   arr2.fill(2.0f);
